mac-addresses: Stop filtering out DVT MACs
We're planning on re-using the DVT range of MAC addresses publicly. This means
the allocation scheme must be the same, and this script must use the same
technique. Since existing DVT units will not be shipped as launched products, we
can safely remap these MACs, despite the collisions.

Sorry, internal devs. :D

Change-Id: I2aaa7d350e85b9a4a5dcaabf0809aa46adee5d7e